ISAKMP: The modecfg attribute type INTERNAL_IP6_ADDRESS is not decoded correctly
According to RFC7296, section 3.15.1:

"The INTERNAL_IP6_ADDRESS is made up of two fields: the first is a 16-octet IPv6 address, and the second is a one-octet prefix-length as defined in [ADDRIPV6]."

In the code, the type is declared as FT_IPv4 and the field is decoded only if it has 16 bytes. Instead, it should be declared as FT_IPv6 and the first 16 bytes should be added to the tree.

Issue reported by Andrei Cipu

Bug: 11393
Ping-Bug: 11392
Change-Id: I8dbc268e71fd6239dffa5469652345a68e0adc86
Reviewed-on: https://code.wireshark.org/review/9761
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>