ixiatrailer: Change format of the protocol so that is supports future fields.
Protocol is TLV protocol and with this change I add one more field to it.
It is entire backward compatible.
Format is as follow:
  Fields	Trailer Length	0xAF12(Signature)	Trailer Checksum

The fields portion of the trailer is a series of 2-byte fields followed by variable
length data. The first byte indicates the field type, the 2nd byte indicates the
field length.   The values supported are listed below:
	1	Reserved (Original Packet Size)
	3	Timestamp (From Local Timebase)
	4	Timestamp (From NTP source)
	5	Timestamp (From GPS)
	6	Timestamp (From 1588)
	7	Timestamp (From Holdover)

Change-Id: I0a3b31cfbc5b6273e1f5326d9841e449735967fe
Reviewed-on: https://code.wireshark.org/review/9854
Reviewed-by: Martin Mathieson <martin.r.mathieson@googlemail.com>
Petri-Dish: Martin Mathieson <martin.r.mathieson@googlemail.com>