Add proper base prefixes in display filter expressions.
When the user selects an enum in the display filter expression dialog
make sure we add the proper base prefix

Add a Qt UI convenience routine for formatting integers. Because
deprecating QString::sprintf was a well-thought-out decision that will
be regarded highly for decades to come.

Fix search visibility while we're here.

Bug: 11421
Change-Id: I1368bc8f464983793bdc90760cf3b7c18b6bd859
Reviewed-on: https://code.wireshark.org/review/9885
Reviewed-by: Gerald Combs <gerald@wireshark.org>