Cleanup a bunch of 802.11ad stuff.
Make all the multi-byte fields, except for OUIs, little-endian -
including fields that contain bitfields.  Make the bits in the bitfields
little-endian as well.

For the "number of taps" bitfields, interpret the values.

Fix a typo ("pseduo" -> "pseudo").

The length of an OCT MMPDU is 16 bits, not 8 bits, and it's not a text
string, it's an octet string.

Handle the beam refinement UI as a 2-byte field that overlaps with a
4-byte field - *none* of the bitfields align on nice 8-bit boundaries,
so that's the best we can do.

Bug: 11419
Change-Id: Ie2d8b1224783a0bef975dcd66f93ce72b35ba9f4
Reviewed-on: https://code.wireshark.org/review/9887
Reviewed-by: Guy Harris <guy@alum.mit.edu>