Un-reference actions after adding them to an action group.
When created, they have a reference count of 1; when added to an action
group, the reference count is increased as a pointer is held by the
action group.  Release the reference we got, as we're not holding onto
it.

Also, rename "menu_item" to "menu_action", to indicate that it's an action.

Change-Id: I54543dcc2e6c3b341a9f584fe1b6099a134771b3
Reviewed-on: https://code.wireshark.org/review/10049
Reviewed-by: Guy Harris <guy@alum.mit.edu>