Preparation Host Flows: Syntax Line Edit (allow longer filter text)
The "Syntax Line Edit" is based on "QLineEdit". By default, the maximum
permitted length of a QLineEdit text is 32767 (16 bits).

As the "Host Flows" extension uses the Wireshark filter syntax to filter
the data, the filter can be become large. To avoid errors, I increased
the maximum permitted length value to the maximum of quint32.

Change-Id: I756efa76e1fc84449caef8fea368e6e38e776235
Reviewed-on: https://code.wireshark.org/review/10099
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>