Check if -fPIC is needed for Qt5 code to compile
On many systems, Qt5 requires that code compiled against it must compile
with -fPIC.  The preferred place to determine whether this is necessary
is at configure time.  This change adds the auto-tool code to make that
happen and removes the hack put in place as an interim solution.

Bug: 11230
Change-Id: I6e583c67bb67d168a30c0af05a8cd0a070522ba0
Reviewed-on: https://code.wireshark.org/review/10281
Petri-Dish: Jeff Morriss <jeff.morriss.ws@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Jeff Morriss <jeff.morriss.ws@gmail.com>