Clean up scanning to guess the bitrate of an MPEG-2 Transport Stream file.
Don't treat an EOF at the beginning of a frame as an error.

Treat I/O errors as hard errors; treat short reads as an indication that
the file isn't an MPEG-2 Transport Stream file.

Treat the PCR for a given PID not going forward as an indication that
the file isn't an MPEG-2 Transport Stream file.

Bug: 11471
Change-Id: I42b5887049423f8265db9d121d7b5bd388e5b244
Reviewed-on: https://code.wireshark.org/review/10286
Reviewed-by: Guy Harris <guy@alum.mit.edu>