Don't check the radio information when testing for an HT Control header.
I now read 8.2.4.1.10 "Order field" in 802.11-2012 as saying that, in
management and QoS data frames, the Order bit shouldn't be set for
non-HT, non-VHT frames, so we can just test it for those frame types
without bothering to check the radio metadata to see if the frame is an
HT or VHT frame.

This handles cases where the radio metadata isn't complete, e.g. an HT
frame with a radiotap header but no MCS field.

Handle this for *all* QoS data frames when capturing.

Get rid of the "fixed-length link-layer header" stuff; it's not being
used.

Fix a case where we're appending text to a tree item without a space
separating it from the previous text.

Bug: 11351
Change-Id: I980f5b7509603b0c22c297fddc19434c08817913
Reviewed-on: https://code.wireshark.org/review/10288
Reviewed-by: Guy Harris <guy@alum.mit.edu>