dtls,ssl: Fix invalid free in UAT key parsing
uat_esc returns ep_alloc0 memory which is automatically freed before the
next frame dissection and must not be passed to g_free(). Fixed a crash
on OS X and an ASAN report.

Note that this problem is not present on master (1.99.x) as it was
refactored to use g_malloc. On 1.12.x however the required changes would
be too large so go for this simpler approach where the caller is
modified.

Bug: 11443
Change-Id: I51a2421cee07d4567cedd7453aa75b8ce7d0433d
Reviewed-on: https://code.wireshark.org/review/10329
Reviewed-by: Michael Mann <mmann78@netscape.net>
Reviewed-by: Peter Wu <peter@lekensteyn.nl>