Squelch 64-bit-to-32-bit warnings by just casting to int.
Those lengths had better fit in an int if they're added to packet offsets.

(BTW, gsize is the spawn of Satan; it should never be used except when
you're dealing with GLib.  It *should* have just been another name for
size_t, but it's 32 bits on 64-bit Windows, which means it's narrower
than size_t, which causes us some pain with g_snprintf().)

Change-Id: Icd8f0632242303dbea0d80e0dad45b317097daaa
Reviewed-on: https://code.wireshark.org/review/10516
Reviewed-by: Guy Harris <guy@alum.mit.edu>