debian: Enable parallel package build
Debian packages used to build with parallel build enabled when
debian/rules used autotools but with switching to CMake the
parallel build failed sometimes. Now the CMake based build
system seems to be working for parallel builds, too, thus
it seems to be safe to enable parallel building of .debs again.

Change-Id: I79003bf6c4b74640f24d907b763a5cc3da595e68
Reviewed-on: https://code.wireshark.org/review/10657
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>