Fix tests to allow selection of suite 10
The tests only allowed a single character suite selection, which
prevented suite 10 from being run on its own.

Modified test.sh to now require a newline in addition to the input
so that a 2 digit number can be entered.

Also fixed test display to remove illusion that an individual step
could be run.  Only whole suites can be run.

Change-Id: I4dee0ec6a8e1f34fa443a6a0a3f2d52a73146e54
Reviewed-on: https://code.wireshark.org/review/10676
Reviewed-by: Graham Bloice <graham.bloice@trihedral.com>